backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0

# AI (Groq)
groq>=0.4.0

# HTTP
requests>=2.31.0

# Env
python-dotenv>=1.0.0
pydantic>=2.0.0
